requirements.txt
torch>=2.3.0
faiss-cpu>=1.8.0
gradio>=5.0
numpy>=1.26.0
tqdm>=4.66.0
cryptography>=43.0.0
scipy>=1.11.0
scikit-learn>=1.3.0
networkx>=3.0
plotly>=5.17.0
pandas>=2.0.0